requirements.txt
tensorflow
opencv-python
numpy
matplotlib
scikit-learn
flask
pillow
seaborn